Add cache manager
