+++ src/core/Jalasoft.TeamUp.Resumes.Core.Interfaces/ISkillsService.cs
@@ -1,6 +1,5 @@
 ﻿namespace Jalasoft.TeamUp.Resumes.Core.Interfaces
 {
-    using System;
     using Jalasoft.TeamUp.Resumes.Models;
 
     public interface ISkillsService
