+++ src/core/Jalasoft.TeamUp.Resumes.Core/SkillsService.cs
@@ -1,7 +1,5 @@
 ﻿namespace Jalasoft.TeamUp.Resumes.Core
 {
-    using System;
-    using System.Collections.Generic;
     using System.Linq;
     using Jalasoft.TeamUp.Resumes.Core.Interfaces;
     using Jalasoft.TeamUp.Resumes.DAL.Interfaces;
